+++ tests/Parser.hs
@@ -139,6 +139,9 @@ parserTests =
             , shouldParse
                 "collections with type annotations containing imports"
                 "./tests/parser/collectionImportType.dhall"
+            , shouldParse
+                "a parenthesized custom header import"
+                "./tests/parser/parenthesizeUsing.dhall"
             ]
         ]